pyproject.toml
[project]
name = "langchain-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-tavily>=0.2.17",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
